[SIFISC-4560] Augmenté la taille de la colonne EVT_ID de la table AUDIT_LOG pour éviter des overflows sur les numéros d'événements RcPers.
git-svn-id: https://svn.etat-de-vaud.ch/SVN_SOCLE/registre/unireg/trunk@75779 cb3efa0a-b8b0-9f43-a4c8-343d8b7e4362

+++ unireg/web/src/main/java/ch/vd/uniregctb/admin/AuditLogBean.java
@@ -18,7 +18,7 @@ public class AuditLogBean {
 	public static class AuditView {
 		private final Long id;
 		private final Long threadId;
-		private final Integer evenementId;
+		private final Long evenementId;
 		private final Date date;
 		private final String user;
 		private final String message;
@@ -51,7 +51,7 @@ public class AuditLogBean {
 			return threadId;
 		}
 
-		public Integer getEvenementId() {
+		public Long getEvenementId() {
 			return evenementId;
 		}
